corection

+++ common/src/engine/renderer.cc
@@ -55,10 +55,10 @@ void DrawCircle(const float centerX, const float centerY, const float radius,
     const float angle1 = 2 * core::PI * static_cast<float>(i) / nbSegments;
     const float angle2 = 2 * core::PI * static_cast<float>(i + 1) / nbSegments;
 
-    const float x1 = centerX + radius * cos(angle1);
-    const float y1 = centerY + radius * sin(angle1);
-    const float x2 = centerX + radius * cos(angle2);
-    const float y2 = centerY + radius * sin(angle2);
+    const float x1 = centerX + radius * static_cast<float>(cos(angle1));
+    const float y1 = centerY + radius * static_cast<float>(sin(angle1));
+    const float x2 = centerX + radius * static_cast<float>(cos(angle2));
+    const float y2 = centerY + radius * static_cast<float>(sin(angle2));
 
     SDL_Vertex vertices[3];
     vertices[0].position = { centerX, centerY };
